Split the raw actor implementation by responsibility
`raw.rs` had grown to ~3.5k lines covering four unrelated concerns. Move
it to a directory module and partition it, without changing any item body:

- `raw/context.rs` — `RawContext`, `Rejected`, `RawResources`, the event
  queue and the timer store, plus the two inline test modules.
- `raw/definition.rs` — `RawActor`, `RawDef`, `RawOnceDef` and their
  erasure into `RawConstruction`/`RawSpawn`/`RawRunContext`.
- `raw/disposal.rs` — `PanicSlot`, `RawDisposal`, `Contained`,
  `CatchUnwindFuture`.
- `raw/offload.rs` — the shared offload state, `Guard`, `Blocking` and
  `DeadlineElapsed`.

`raw/mod.rs` keeps the module header and re-exports the previous surface,
so nothing outside `raw` changes. Only visibility widens: items a sibling
module now reaches are `pub(super)`, and nothing gains `pub` or
`pub(crate)` that did not already have it.

The `RawContext` doc link is written against the public `crate::RawActor`
path, `std::task::Context` keeps its `TaskPollContext` alias in the two
files that poll futures, and each new file carries a `//!` header.
