+++ crates/shelterwood/src/raw/context.rs
@@ -1,4 +1,4 @@
-//! Minimal loop-owning raw actors and their incarnation context.
+//! Per-incarnation raw actor context and owned resources.
 
 use std::{
     any::{Any, TypeId},
@@ -6,41 +6,34 @@ use std::{
     fmt,
     future::Future,
     hash::{BuildHasher, Hash},
-    pin::Pin,
     sync::{Arc, Mutex},
-    task::{Context as TaskPollContext, Poll},
     time::{Duration, Instant},
 };
 
 use crate::{
-    ActorRef, ChildId, DeadlineBudget, ExitResult, Incarnation, Mailbox, MailboxShutdown,
-    PolicyError, Readiness, ReadinessDeadline, RestartPolicy, Retention, Shutdown,
+    ActorRef, ChildId, DeadlineBudget, Incarnation, MailboxShutdown, Readiness,
     cancellation::{CancellationToken, ParentCancellationToken},
-    cells::MemberCell,
-    definition::DefinitionSource,
     identity::PoisonedCounter,
-    mailbox::{
-        AcceptedSequence, MailboxCell, MailboxControl, MailboxEffectQueue, MailboxReceiver,
-        actor_ref_from_parts,
-    },
-    policy::{ChildMode, CommonOptions},
+    mailbox::{AcceptedSequence, MailboxCell, MailboxReceiver},
     runtime::{
-        self, ActorWork, CompletionGatedLatch, Latch, PanicAccumulator, PanicPayload, Signal,
-        SignalWatcher, UnwindPanics, catch_panic, discard_panic, keep_first_panic,
-        resume_preferred_panic, resume_preferred_panic_outside_unwind,
+        self, CompletionGatedLatch, Latch, PanicAccumulator, PanicPayload, Signal, SignalWatcher,
+        UnwindPanics, catch_panic, resume_preferred_panic_outside_unwind,
     },
     scope::ScopeRef,
 };
 
+use super::{
+    definition::RawRunContext,
+    disposal::{CatchUnwindFuture, Contained, PanicSlot, RawDisposal},
+    offload::{
+        Blocking, DeadlineElapsed, Guard, OffloadResource, SharedOffloadFuture, SharedOffloadState,
+    },
+};
+
+#[cfg(test)]
+use super::offload::OffloadPoll;
+
 type DeferredMessage<M> = Box<dyn FnOnce() -> M + Send + 'static>;
-type OffloadFuture = Pin<Box<dyn Future<Output = ()> + Send + 'static>>;
-type SharedWork = Arc<SharedOffloadState>;
-
-/// Marker returned to an offload continuation when its one deadline expires.
-#[derive(Clone, Copy, Debug, Eq, PartialEq, thiserror::Error)]
-#[error("the offload deadline elapsed")]
-pub struct DeadlineElapsed;
-
 /// An operation rejected because the actor incarnation is already stopping.
 #[derive(Eq, PartialEq)]
 pub struct Rejected<T> {
@@ -65,275 +58,9 @@ impl<T> Rejected<T> {
     }
 }
 
-/// An owned cancel-on-drop lease for a scoped offload.
-#[must_use = "dropping the guard cancels its offload; call detach to keep only incarnation ownership"]
-pub struct Guard {
-    cancellation: Latch,
-    finished: Latch,
-    armed: bool,
-}
-
-impl fmt::Debug for Guard {
-    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-        formatter
-            .debug_struct("Guard")
-            .field("detached", &!self.armed)
-            .finish()
-    }
-}
-
-impl Guard {
-    /// Reports whether cancellation has been requested for this lease.
-    #[must_use]
-    pub fn is_cancelled(&self) -> bool {
-        self.cancellation.is_fired()
-    }
-
-    /// Reports whether the work completed or incarnation teardown requested cancellation.
-    ///
-    /// A teardown notification is not a join: under hard abort the task may
-    /// still be unwinding when this becomes true.
-    #[must_use]
-    pub fn is_finished(&self) -> bool {
-        self.finished.is_fired()
-    }
-
-    /// Waits for work completion or an incarnation-teardown cancellation request.
-    ///
-    /// This notification does not join work that is being hard-aborted.
-    pub async fn finished(&self) {
-        self.finished.fired().await;
-    }
-
-    /// Cancels the guarded offload immediately and consumes the guard.
-    pub fn cancel(mut self) {
-        self.cancellation.fire();
-        self.armed = false;
-    }
-
-    /// Releases this lease's cancel-on-drop behavior.
-    pub fn detach(mut self) {
-        self.armed = false;
-    }
-}
-
-impl Drop for Guard {
-    fn drop(&mut self) {
-        if self.armed {
-            self.cancellation.fire();
-        }
-    }
-}
-
-/// A blocking operation whose thread cooperatively observes actor cancellation.
-///
-/// Cancellation cannot forcibly stop a blocking thread. Dropping this future
-/// or hard-aborting its actor detaches the thread after requesting cooperative
-/// cancellation; Shelterwood does not join it, and any later value or panic is
-/// discarded through detached disposal. A submission rejected during runtime
-/// teardown moves to a detached Shelterwood thread; an operation that never
-/// runs — cancelled with the runtime, or with no thread left to start it —
-/// makes awaiting the future panic with a runtime-teardown cancellation
-/// diagnostic. The operation must therefore be safe to outlive its actor.
-#[must_use = "dropping this future requests cooperative cancellation and detaches the thread"]
-pub struct Blocking<T> {
-    future: Pin<Box<dyn Future<Output = T> + Send + 'static>>,
-    cancellation: Latch,
-    completed: bool,
-}
-
-impl<T> fmt::Debug for Blocking<T> {
-    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-        formatter
-            .debug_struct("Blocking")
-            .field("completed", &self.completed)
-            .finish_non_exhaustive()
-    }
-}
-
-impl<T> Future for Blocking<T> {
-    type Output = T;
-
-    fn poll(mut self: Pin<&mut Self>, context: &mut TaskPollContext<'_>) -> Poll<Self::Output> {
-        match self.future.as_mut().poll(context) {
-            Poll::Ready(value) => {
-                self.completed = true;
-                Poll::Ready(value)
-            }
-            Poll::Pending => Poll::Pending,
-        }
-    }
-}
-
-impl<T> Drop for Blocking<T> {
-    fn drop(&mut self) {
-        if !self.completed {
-            self.cancellation.fire();
-        }
-    }
-}
-
-/// Future panic boundary that owns and destroys its inner future.
-///
-/// Once the inner future returns `Ready`, it is destroyed before the output is
-/// released. If that destruction panics, the already-produced output is
-/// discarded and the destructor panic becomes this future's error.
-pub(crate) struct CatchUnwindFuture<F> {
-    future: Option<Pin<Box<F>>>,
-}
-
-impl<F> CatchUnwindFuture<F> {
-    pub(crate) fn new(future: F) -> Self {
-        Self {
-            future: Some(Box::pin(future)),
-        }
-    }
-}
-
-impl<F: Future> Future for CatchUnwindFuture<F> {
-    type Output = Result<F::Output, PanicPayload>;
-
-    fn poll(self: Pin<&mut Self>, context: &mut TaskPollContext<'_>) -> Poll<Self::Output> {
-        let this = self.get_mut();
-        let polled = catch_panic(|| {
-            this.future
-                .as_mut()
-                .expect("a completed panic boundary was polled again")
-                .as_mut()
-                .poll(context)
-        });
-        match polled {
-            Ok(Poll::Ready(value)) => {
-                let future = this.future.take();
-                match catch_panic(|| drop(future)) {
-                    Ok(()) => Poll::Ready(Ok(value)),
-                    Err(payload) => Poll::Ready(Err(payload)),
-                }
-            }
-            Ok(Poll::Pending) => Poll::Pending,
-            Err(payload) => {
-                let future = this.future.take();
-                discard_panic(catch_panic(|| drop(future)).err());
-                Poll::Ready(Err(payload))
-            }
-        }
-    }
-}
-
-impl<F> Drop for CatchUnwindFuture<F> {
-    fn drop(&mut self) {
-        let future = self.future.take();
-        let mut panics = PanicAccumulator::default();
-        panics.run(|| drop(future));
-    }
-}
-
 struct QueuedEvent<M> {
     cancellation: Latch,
     make_message: DeferredMessage<M>,
-}
-
-#[derive(Default)]
-struct PanicSlot {
-    payload: Mutex<Option<PanicPayload>>,
-}
-
-impl PanicSlot {
-    fn record(&self, payload: PanicPayload) {
-        let rejected = {
-            let mut pending = self.payload.lock().expect("offload panic mutex poisoned");
-            if pending.is_none() {
-                *pending = Some(payload);
-                None
-            } else {
-                Some(payload)
-            }
-        };
-        discard_panic(rejected);
-    }
-
-    fn take(&self) -> Option<PanicPayload> {
-        self.payload
-            .lock()
-            .expect("offload panic mutex poisoned")
-            .take()
-    }
-}
-
-/// The one cleanup route for values owned by a raw incarnation.
-///
-/// Collection drains and offload futures route user payloads through this
-/// funnel. Collections keep their resident elements raw and dispose each one
-/// explicitly when draining, avoiding a cloned disposal handle per element.
-/// A destructor panic is retained as cleanup evidence and wakes an idle actor;
-/// it never unwinds through another user destructor.
-#[derive(Clone)]
-struct RawDisposal {
-    panic: Arc<PanicSlot>,
-    signal: Signal,
-}
-
-/// Test-only: mints an orphan disposal whose panic slot and signal nothing
-/// observes. Production wiring threads one shared disposal per incarnation
-/// through the container constructors.
-#[cfg(test)]
-impl Default for RawDisposal {
-    fn default() -> Self {
-        Self {
-            panic: Arc::new(PanicSlot::default()),
-            signal: Signal::default(),
-        }
-    }
-}
-
-impl RawDisposal {
-    fn record(&self, payload: PanicPayload) {
-        self.panic.record(payload);
-        if let Err(payload) = catch_panic(|| self.signal.pulse()) {
-            self.panic.record(payload);
-        }
-    }
-
-    fn dispose<T>(&self, value: T) {
-        if let Err(payload) = catch_panic(|| drop(value)) {
-            self.record(payload);
-        }
-    }
-}
-
-#[must_use = "contained user ownership must be consumed or disposed"]
-struct Contained<T> {
-    value: Option<T>,
-    disposal: RawDisposal,
-}
-
-impl<T> Contained<T> {
-    fn new(value: T, disposal: RawDisposal) -> Self {
-        Self {
-            value: Some(value),
-            disposal,
-        }
-    }
-
-    fn get(&self) -> &T {
-        self.value
-            .as_ref()
-            .expect("contained ownership is consumed once")
-    }
-
-    fn into_inner(mut self) -> T {
-        self.value
-            .take()
-            .expect("contained ownership is consumed once")
-    }
-}
-
-impl<T> Drop for Contained<T> {
-    fn drop(&mut self) {
-        if let Some(value) = self.value.take() {
-            self.disposal.dispose(value);
-        }
-    }
 }
 
 /// Incarnation-internal completion storage for offload continuations.
@@ -930,154 +657,6 @@ impl ReadyBatch {
     }
 }
 
-struct OffloadFutureState {
-    future: Option<OffloadFuture>,
-    polling: bool,
-    cancelled: bool,
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-enum OffloadPoll {
-    Pending,
-    Finished,
-}
-
-struct SharedOffloadState {
-    // Polling takes the future out of this mutex. Cancellation either takes
-    // an idle future or marks an in-progress poll so that the poller disposes
-    // it, always after releasing the lock.
-    state: Mutex<OffloadFutureState>,
-    disposal: RawDisposal,
-    finished: Latch,
-}
-
-impl SharedOffloadState {
-    fn new(future: OffloadFuture, disposal: RawDisposal, finished: Latch) -> SharedWork {
-        Arc::new(Self {
-            state: Mutex::new(OffloadFutureState {
-                future: Some(future),
-                polling: false,
-                cancelled: false,
-            }),
-            disposal,
-            finished,
-        })
-    }
-
-    fn take_for_poll(&self) -> Option<OffloadFuture> {
-        let mut state = self.state.lock().expect("offload future mutex poisoned");
-        if state.cancelled {
-            return None;
-        }
-        debug_assert!(!state.polling, "offload work must have one poller");
-        let future = state.future.take();
-        state.polling = future.is_some();
-        future
-    }
-
-    fn finish_poll(&self, future: OffloadFuture, outcome: OffloadPoll) -> Option<OffloadFuture> {
-        let mut state = self.state.lock().expect("offload future mutex poisoned");
-        state.polling = false;
-        if outcome == OffloadPoll::Pending && !state.cancelled {
-            debug_assert!(state.future.is_none());
-            state.future = Some(future);
-            None
-        } else {
-            Some(future)
-        }
-    }
-
-    fn record(&self, payload: PanicPayload) {
-        // Dropping a losing or cancelled operation can panic after its body
-        // has stopped running, so every retained panic must wake the actor's
-        // control plane independently of ordinary event delivery.
-        self.disposal.record(payload);
-    }
-
-    fn dispose(&self, future: Option<OffloadFuture>) {
-        if let Some(future) = future {
-            self.disposal.dispose(future);
-        }
-    }
-
-    fn cancel(&self) {
-        let future = {
-            let mut state = self.state.lock().expect("offload future mutex poisoned");
-            state.cancelled = true;
-            if state.polling {
-                None
-            } else {
-                state.future.take()
-            }
-        };
-        self.dispose(future);
-        self.finished.fire();
-    }
-}
-
-struct SharedOffloadFuture(SharedWork);
-
-impl Future for SharedOffloadFuture {
-    type Output = ();
-
-    fn poll(self: Pin<&mut Self>, context: &mut TaskPollContext<'_>) -> Poll<Self::Output> {
-        let Some(mut future) = self.0.take_for_poll() else {
-            return Poll::Ready(());
-        };
-        let polled = catch_panic(|| future.as_mut().poll(context));
-        match polled {
-            Ok(Poll::Pending) => {
-                let dispose = self.0.finish_poll(future, OffloadPoll::Pending);
-                if dispose.is_some() {
-                    self.0.dispose(dispose);
-                    self.0.finished.fire();
-                    Poll::Ready(())
-                } else {
-                    Poll::Pending
-                }
-            }
-            Ok(Poll::Ready(())) => {
-                let dispose = self.0.finish_poll(future, OffloadPoll::Finished);
-                self.0.dispose(dispose);
-                self.0.finished.fire();
-                Poll::Ready(())
-            }
-            Err(payload) => {
-                let dispose = self.0.finish_poll(future, OffloadPoll::Finished);
-                self.0.record(payload);
-                self.0.dispose(dispose);
-                self.0.finished.fire();
-                Poll::Ready(())
-            }
-        }
-    }
-}
-
-struct OffloadResource {
-    cancellation: Latch,
-    finished: Latch,
-    state: Option<SharedWork>,
-    task: Option<ActorWork>,
-}
-
-impl OffloadResource {
-    fn cancel(&mut self) {
-        self.cancellation.fire();
-        if let Some(state) = &self.state {
-            state.cancel();
-        }
-        if let Some(task) = &self.task {
-            // These are complementary: `state.cancel()` synchronously
-            // disposes idle work (capturing destructor panic) or marks an
-            // in-progress poll to dispose on return, while abort independently
-            // requests cancellation of the runtime task driving that poll.
-            // Neither substitutes for the other.
-            task.abort();
-        }
-        self.finished.fire();
-    }
-}
-
 struct RawResources<M> {
     accepting: bool,
     continuations: ContinuationQueue<M>,
@@ -1195,40 +774,7 @@ impl<M> Drop for RawResources<M> {
     }
 }
 
-/// Minimal actor contract for application-owned receive loops.
-pub trait RawActor: Send + 'static {
-    /// Message accepted by this actor.
-    type Msg: Send + 'static;
-
-    /// Declares when this actor type becomes ready.
-    ///
-    /// This is definition metadata: the framework reads it before constructing
-    /// an incarnation, so it cannot depend on per-incarnation actor state.
-    fn readiness() -> Readiness {
-        Readiness::Immediate
-    }
-
-    /// Runs one incarnation using the membership-owned mailbox binding.
-    ///
-    /// The framework calls this method at most once on an incarnation's root
-    /// raw-actor value and never re-enters it on that value. Shutdown may
-    /// destroy a constructed root before its run begins; a restart that reaches
-    /// construction obtains a fresh root value.
-    ///
-    /// [`RawContext::recv`] freezes external intake and returns `None` when
-    /// shutdown begins. A raw loop must then honor
-    /// [`RawContext::mailbox_shutdown`]: for
-    /// [`MailboxShutdown::Drain`], repeatedly call [`RawContext::try_recv`] to
-    /// handle the frozen accepted prefix; for [`MailboxShutdown::Discard`],
-    /// return without draining. The high-level [`crate::Actor`] loop implements
-    /// this policy automatically.
-    fn run(
-        &mut self,
-        context: &mut RawContext<Self::Msg>,
-    ) -> impl Future<Output = ExitResult> + Send;
-}
-
-/// Per-incarnation capabilities supplied to a [`RawActor`].
+/// Per-incarnation capabilities supplied to a [`RawActor`](crate::RawActor).
 pub struct RawContext<M> {
     id: ChildId,
     incarnation: Incarnation,
@@ -1255,7 +801,7 @@ impl<M> fmt::Debug for RawContext<M> {
 }
 
 impl<M: Send + 'static> RawContext<M> {
-    fn new(
+    pub(super) fn new(
         run: RawRunContext,
         myself: ActorRef<M>,
         mailbox: Arc<MailboxCell<M>>,
@@ -1971,314 +1517,9 @@ impl<M: Send + 'static> RawContext<M> {
         self.resources.join_offloads().await;
     }
 
-    fn take_resource_panic(&self) -> Option<PanicPayload> {
+    pub(super) fn take_resource_panic(&self) -> Option<PanicPayload> {
         self.resources.disposal.panic.take()
     }
-}
-
-/// Restartable raw-actor definition.
-pub struct RawDef<R: RawActor> {
-    factory: Box<dyn Fn() -> R + Send + Sync + 'static>,
-    pub(crate) options: CommonOptions,
-}
-
-impl<R: RawActor> fmt::Debug for RawDef<R> {
-    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-        formatter
-            .debug_struct("RawDef")
-            .field("options", &self.options)
-            .finish_non_exhaustive()
-    }
-}
-
-impl<R: RawActor> RawDef<R> {
-    /// Creates a restartable definition from a repeatable actor factory.
-    pub fn factory(factory: impl Fn() -> R + Send + Sync + 'static) -> Self {
-        Self {
-            factory: Box::new(factory),
-            options: CommonOptions::default(),
-        }
-    }
-
-    common_options_setters!(
-        restart,
-        shutdown,
-        mailbox,
-        mailbox_shutdown,
-        raw_readiness,
-        structural_readiness_deadline,
-        retention,
-    );
-
-    pub(crate) fn erase(self, mailbox: Arc<MailboxCell<R::Msg>>) -> RawConstruction {
-        let factory = self.factory;
-        let readiness = self.options.readiness.unwrap_or_else(R::readiness);
-        RawConstruction {
-            source: DefinitionSource::Restartable(Arc::new(move || {
-                let actor = factory();
-                Box::new(RawInstance {
-                    actor,
-                    mailbox: Arc::clone(&mailbox),
-                })
-            })),
-            options: self.options,
-            readiness,
-        }
-    }
-}
-
-/// Consuming one-shot raw-actor definition.
-pub struct RawOnceDef<R: RawActor> {
-    actor: R,
-    pub(crate) options: CommonOptions,
-}
-
-impl<R: RawActor> fmt::Debug for RawOnceDef<R> {
-    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-        formatter
-            .debug_struct("RawOnceDef")
-            .field("actor", &"<owned raw actor>")
-            .field("options", &self.options)
-            .finish()
-    }
-}
-
-impl<R: RawActor> RawOnceDef<R> {
-    /// Creates a one-shot definition from an owned actor value.
-    pub fn new(actor: R) -> Self {
-        Self {
-            actor,
-            options: CommonOptions::default(),
-        }
-    }
-
-    common_options_setters!(
-        shutdown,
-        mailbox,
-        mailbox_shutdown,
-        raw_readiness,
-        structural_readiness_deadline,
-        retention,
-    );
-
-    pub(crate) fn erase(self, mailbox: Arc<MailboxCell<R::Msg>>) -> RawConstruction {
-        let readiness = self.options.readiness.unwrap_or_else(R::readiness);
-        RawConstruction {
-            source: DefinitionSource::OneShot(Box::new(RawInstance {
-                actor: self.actor,
-                mailbox,
-            })),
-            options: self.options,
-            readiness,
-        }
-    }
-}
-
-type RawFuture = Pin<Box<dyn Future<Output = ExitResult> + Send + 'static>>;
-type RawFactory = Arc<dyn Fn() -> Box<dyn ErasedRawInstance> + Send + Sync + 'static>;
-
-trait ErasedRawInstance: Send {
-    fn run(self: Box<Self>, context: RawRunContext, readiness: Readiness) -> RawFuture;
-}
-
-struct RawInstance<R: RawActor> {
-    actor: R,
-    mailbox: Arc<MailboxCell<R::Msg>>,
-}
-
-struct RawIncarnationOwner<R: RawActor> {
-    raw: Option<RawContext<R::Msg>>,
-    actor: Option<R>,
-    primary_panic: Option<PanicPayload>,
-}
-
-impl<R: RawActor> RawIncarnationOwner<R> {
-    fn new(raw: RawContext<R::Msg>, actor: R) -> Self {
-        Self {
-            raw: Some(raw),
-            actor: Some(actor),
-            primary_panic: None,
-        }
-    }
-
-    fn parts(&mut self) -> (&mut R, &mut RawContext<R::Msg>) {
-        let actor = self.actor.as_mut().expect("raw actor owner is armed");
-        let raw = self.raw.as_mut().expect("raw context owner is armed");
-        (actor, raw)
-    }
-
-    fn raw(&mut self) -> &mut RawContext<R::Msg> {
-        self.raw.as_mut().expect("raw context owner is armed")
-    }
-
-    fn drop_raw(&mut self) {
-        drop(self.raw.take());
-    }
-
-    fn drop_actor(&mut self) {
-        drop(self.actor.take());
-    }
-
-    fn record_primary_panic(&mut self, payload: PanicPayload) {
-        debug_assert!(self.primary_panic.is_none());
-        self.primary_panic = Some(payload);
-    }
-
-    fn take_primary_panic(&mut self) -> Option<PanicPayload> {
-        self.primary_panic.take()
-    }
-}
-
-impl<R: RawActor> Drop for RawIncarnationOwner<R> {
-    fn drop(&mut self) {
-        // A hard abort destroys the incarnation future instead of polling its
-        // teardown epilogue. Preserve §5.5's resource-before-actor order, but
-        // put a boundary around each destructor so two panics cannot abort the
-        // process. The resource panic is primary: it may be an owned offload
-        // panic that completed before cancellation was requested.
-        let primary_panic = self.take_primary_panic();
-        let mut cleanup = PanicAccumulator::default();
-        cleanup.run(|| self.drop_raw());
-        cleanup.run(|| self.drop_actor());
-        resume_preferred_panic(UnwindPanics {
-            primary: primary_panic,
-            cleanup: cleanup.take(),
-        });
-    }
-}
-
-impl<R: RawActor> ErasedRawInstance for RawInstance<R> {
-    fn run(self: Box<Self>, context: RawRunContext, readiness: Readiness) -> RawFuture {
-        Box::pin(async move {
-            let Self { actor, mailbox } = *self;
-            let incarnation = context.incarnation;
-            let myself = actor_ref_from_parts(Arc::clone(&context.member), Arc::clone(&mailbox));
-            let raw = RawContext::new(context, myself, Arc::clone(&mailbox), readiness);
-            let mut owner = RawIncarnationOwner::new(raw, actor);
-            let outcome = {
-                let (actor, raw) = owner.parts();
-                CatchUnwindFuture::new(actor.run(raw)).await
-            };
-            let result = match outcome {
-                Ok(result) => Some(result),
-                Err(payload) => {
-                    // Keep the actor's diagnostic in the owned epilogue so a
-                    // hard abort during async teardown cannot replace it with
-                    // a later cancellation or destructor panic.
-                    owner.record_primary_panic(payload);
-                    None
-                }
-            };
-            let mailbox_freeze_panic = catch_panic(|| {
-                let mut effects = MailboxEffectQueue::default();
-                mailbox.freeze(incarnation, &mut effects);
-            })
-            .err();
-            let resource_freeze_panic = catch_panic(|| owner.raw().freeze_resources()).err();
-            let mut cleanup_panic = owner.raw().take_resource_panic();
-            keep_first_panic(&mut cleanup_panic, mailbox_freeze_panic);
-            keep_first_panic(&mut cleanup_panic, resource_freeze_panic);
-
-            let joined = CatchUnwindFuture::new(owner.raw().join_resources()).await;
-            keep_first_panic(&mut cleanup_panic, joined.err());
-            let pending = owner.raw().take_resource_panic();
-            keep_first_panic(&mut cleanup_panic, pending);
-            let raw_drop = catch_panic(|| owner.drop_raw()).err();
-            keep_first_panic(&mut cleanup_panic, raw_drop);
-
-            let actor_drop = catch_panic(|| owner.drop_actor()).err();
-            keep_first_panic(&mut cleanup_panic, actor_drop);
-            // Once actor execution has panicked, teardown is secondary: never
-            // replace the actor's original diagnostic. This is the incarnation
-            // body's normal return path, so the resume must be unconditional:
-            // containing the primary payload here would strand `result` at
-            // `None` and report the actor's panic as the framework expect
-            // below.
-            resume_preferred_panic_outside_unwind(UnwindPanics {
-                primary: owner.take_primary_panic(),
-                cleanup: cleanup_panic,
-            });
-            result.expect("an incarnation without a primary panic returns a result")
-        })
-    }
-}
-
-pub(crate) struct RawConstruction {
-    source: DefinitionSource<RawFactory, Box<dyn ErasedRawInstance>>,
-    options: CommonOptions,
-    readiness: Readiness,
-}
-
-impl RawConstruction {
-    pub(crate) fn options(&self) -> &CommonOptions {
-        &self.options
-    }
-
-    pub(crate) fn readiness(&self) -> Readiness {
-        self.readiness
-    }
-
-    pub(crate) fn mode(&self) -> ChildMode {
-        if self.source.is_one_shot() {
-            ChildMode::OneShot
-        } else {
-            ChildMode::Restartable
-        }
-    }
-
-    pub(crate) fn one_shot(&self) -> bool {
-        self.source.is_one_shot()
-    }
-
-    pub(crate) fn take_spawn(&mut self) -> RawSpawn {
-        if let Some(factory) = self.source.restartable() {
-            RawSpawn(RawSpawnKind::Restartable(Arc::clone(factory)))
-        } else {
-            RawSpawn(RawSpawnKind::OneShot(self.source.take_one_shot().expect(
-                "one-shot raw actor construction invoked more than once",
-            )))
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn for_policy_test(options: CommonOptions, readiness: Readiness) -> Self {
-        Self {
-            source: DefinitionSource::Restartable(Arc::new(|| {
-                unreachable!("policy resolution never constructs the actor")
-            })),
-            options,
-            readiness,
-        }
-    }
-}
-
-pub(crate) struct RawSpawn(RawSpawnKind);
-
-enum RawSpawnKind {
-    Restartable(RawFactory),
-    OneShot(Box<dyn ErasedRawInstance>),
-}
-
-impl RawSpawn {
-    pub(crate) async fn run(self, context: RawRunContext, readiness: Readiness) -> ExitResult {
-        let instance = match self.0 {
-            RawSpawnKind::Restartable(factory) => factory(),
-            RawSpawnKind::OneShot(instance) => instance,
-        };
-        instance.run(context, readiness).await
-    }
-}
-
-pub(crate) struct RawRunContext {
-    pub(crate) id: ChildId,
-    pub(crate) incarnation: Incarnation,
-    pub(crate) member: Arc<MemberCell>,
-    pub(crate) scope: ScopeRef,
-    pub(crate) shutdown: Latch,
-    pub(crate) abort: Latch,
-    pub(crate) ready: CompletionGatedLatch,
-    pub(crate) local_stop: Latch,
-    pub(crate) mailbox_shutdown: MailboxShutdown,
 }
 
 #[cfg(test)]
