+++ crates/shelterwood/src/raw/mod.rs
@@ -1,0 +1,13 @@
+//! Minimal loop-owning raw actors and their incarnation context.
+
+mod context;
+mod definition;
+mod disposal;
+mod offload;
+
+pub use context::{RawContext, Rejected};
+pub use definition::{RawActor, RawDef, RawOnceDef};
+pub use offload::{Blocking, DeadlineElapsed, Guard};
+
+pub(crate) use definition::{RawConstruction, RawRunContext, RawSpawn};
+pub(crate) use disposal::CatchUnwindFuture;
